refactor all settings and services used by viscel into its own class
uses lazy vals for initialization order. the basic idea is that
everything that uses something just reads the value, so starts the
service kinda on demand. in most cases everything will be started as
usual, but now we do not have to care so much about the order in which
things are started.